Working RL code for 3,5 floor building.
Can we add a reset function to the simulation to reset all the states? At the beginning of learning, the elevator can't keep up with how fast people show up so the states it's learning change too fast. Being able to reset after learning some would be a big improvement.

Also, the elevator might be too slow for 10 floors with the given rate for the Poisson process, but idk.

This bug also happened where the elevator wouldn't pick the person up and it also wouldn't leave the floor because it wanted to pick the person up.